test fp 16

+++ src/models/components/nets/new_matcher/new_matcher_net.py
@@ -185,7 +185,8 @@ class NewMatcherNet(nn.Module):
 
         (s1, s2), w = self.scales, self.reg_w
         grid = K.create_meshgrid(
-            s1, s1, normalized_coordinates=False, device=x0_reg.device)
+            s1, s1, normalized_coordinates=False, device=x0_reg.device,
+            dtype=x0_reg.dtype)
         grid = (2 * (grid + 0.5) / s1 - 1).expand(2 * len(x0_reg), -1, -1, -1)
         x = torch.cat([x0_reg, x1_reg]).transpose(1, 2).unflatten(2, (w, w))
         x = F.grid_sample(x, grid, mode="bilinear", align_corners=True)
